Update meter_page : Seperated meter_control widget
